同步

+++ src/component/sheet.js
@@ -5,10 +5,12 @@ import Resizer from './resizer';
 import Scrollbar from './scrollbar';
 import Selector from './selector';
 import Editor from './editor';
+import Print from './print';
 import ContextMenu from './contextmenu';
 import Table from './table';
-import Toolbar from './toolbar';
+import Toolbar from './toolbar/index';
 import ModalValidation from './modal_validation';
+import SortFilter from './sort_filter';
 import { xtoast } from './message';
 import { cssPrefix } from '../config';
 import { formulas } from '../core/formula';
@@ -41,14 +43,14 @@ function scrollbarMove() {
   }
 }
 
-function selectorSet(multiple, ri, ci, indexesUpdated = true) {
+function selectorSet(multiple, ri, ci, indexesUpdated = true, moving = false) {
   if (ri === -1 && ci === -1) return;
   // console.log(multiple, ', ri:', ri, ', ci:', ci);
   const {
     table, selector, toolbar,
   } = this;
   if (multiple) {
-    selector.setEnd(ri, ci);
+    selector.setEnd(ri, ci, moving);
   } else {
     selector.set(ri, ci, indexesUpdated);
   }
@@ -131,13 +133,18 @@ function overlayerMousemove(evt) {
 }
 
 function overlayerMousescroll(evt) {
-  const { verticalScrollbar, data } = this;
+  const { verticalScrollbar, horizontalScrollbar, data } = this;
   const { top } = verticalScrollbar.scroll();
+  const { left } = horizontalScrollbar.scroll();
   // console.log('evt:::', evt.wheelDelta, evt.detail * 40);
-  let delta = evt.deltaY;
-  const { rows } = data;
-  if (evt.detail) delta = evt.detail * 40;
-  if (delta > 0) {
+
+  const { rows, cols } = data;
+
+  // deltaY for vertical delta
+  let { deltaY } = evt;
+  // console.log('deltaX', deltaX, 'evt.detail', evt.detail);
+  if (evt.detail) deltaY = evt.detail * 40;
+  if (deltaY > 0) {
     // up
     const ri = data.scroll.ri + 1;
     if (ri < rows.len) {
@@ -150,13 +157,31 @@ function overlayerMousescroll(evt) {
       verticalScrollbar.move({ top: ri === 0 ? 0 : top - rows.getHeight(ri) });
     }
   }
+
+  // deltaX for Mac horizontal scroll
+  const { deltaX } = evt;
+  if (deltaX > 0) {
+    // left
+    const ci = data.scroll.ci + 1;
+    if (ci < cols.len) {
+      horizontalScrollbar.move({ left: left + cols.getWidth(ci) - 1 });
+    }
+  } else {
+    // right
+    const ci = data.scroll.ci - 1;
+    if (ci >= 0) {
+      horizontalScrollbar.move({
+        left: ci === 0 ? 0 : left - cols.getWidth(ci),
+      });
+    }
+  }
 }
 
 function overlayerTouch(direction, distance) {
   const { verticalScrollbar, horizontalScrollbar } = this;
   const { top } = verticalScrollbar.scroll();
   const { left } = horizontalScrollbar.scroll();
-  // console.log('direction:', direction, ', distance:', distance, left);
+
   if (direction === 'left' || direction === 'right') {
     horizontalScrollbar.move({ left: left - distance });
   } else if (direction === 'up' || direction === 'down') {
@@ -167,7 +192,9 @@ function overlayerTouch(direction, distance) {
 function verticalScrollbarSet() {
   const { data, verticalScrollbar } = this;
   const { height } = this.getTableOffset();
-  verticalScrollbar.set(height, data.rows.totalHeight());
+  const erth = data.exceptRowTotalHeight(0, -1);
+  // console.log('erth:', erth);
+  verticalScrollbar.set(height, data.rows.totalHeight() - erth);
 }
 
 function horizontalScrollbarSet() {
@@ -240,6 +267,12 @@ function paste(what) {
   }
 }
 
+function autofilter() {
+  const { data } = this;
+  data.autofilter();
+  sheetReset.call(this);
+}
+
 function toolbarChangePaintformatPaste() {
   const { toolbar } = this;
   if (toolbar.paintformatActive()) {
@@ -253,10 +286,26 @@ function overlayerMousedown(evt) {
   // console.log(':::::overlayer.mousedown:', evt.detail, evt.button, evt.buttons, evt.shiftKey);
   // console.log('evt.target.className:', evt.target.className);
   const {
-    selector, data, table,
+    selector, data, table, sortFilter,
   } = this;
+  const { offsetX, offsetY } = evt;
   const isAutofillEl = evt.target.className === `${cssPrefix}-selector-corner`;
-  let { ri, ci } = data.getCellRectByXY(evt.offsetX, evt.offsetY);
+  const cellRect = data.getCellRectByXY(offsetX, offsetY);
+  const {
+    left, top, width, height,
+  } = cellRect;
+  let { ri, ci } = cellRect;
+  // sort or filter
+  const { autoFilter } = data;
+  if (autoFilter.includes(ri, ci)) {
+    if (left + width - 20 < offsetX && top + height - 20 < offsetY) {
+      const items = autoFilter.items(ci, (r, c) => data.rows.getCell(r, c));
+      sortFilter.set(ci, items, autoFilter.getFilter(ci), autoFilter.getSort(ci));
+      sortFilter.setOffset({ left, top: top + height + 2 });
+      return;
+    }
+  }
+
   // console.log('ri:', ri, ', ci:', ci);
   if (!evt.shiftKey) {
     // console.log('selectorSetStart:::');
@@ -273,7 +322,7 @@ function overlayerMousedown(evt) {
       if (isAutofillEl) {
         selector.showAutofill(ri, ci);
       } else if (e.buttons === 1 && !e.shiftKey) {
-        selectorSet.call(this, true, ri, ci);
+        selectorSet.call(this, true, ri, ci, true, true);
       }
     }, () => {
       if (isAutofillEl) {
@@ -375,6 +424,14 @@ function insertDeleteRowColumn(type) {
     data.deleteCell('format');
   } else if (type === 'delete-cell-text') {
     data.deleteCell('text');
+  } else if (type === 'cell-printable') {
+    data.setSelectedCellAttr('printable', true);
+  } else if (type === 'cell-non-printable') {
+    data.setSelectedCellAttr('printable', false);
+  } else if (type === 'cell-editable') {
+    data.setSelectedCellAttr('editable', true);
+  } else if (type === 'cell-non-editable') {
+    data.setSelectedCellAttr('editable', false);
   }
   clearClipboard.call(this);
   sheetReset.call(this);
@@ -387,7 +444,7 @@ function toolbarChange(type, value) {
   } else if (type === 'redo') {
     this.redo();
   } else if (type === 'print') {
-    // print
+    this.print.preview();
   } else if (type === 'paintformat') {
     if (value === true) copy.call(this);
     else clearClipboard.call(this);
@@ -397,8 +454,9 @@ function toolbarChange(type, value) {
     // link
   } else if (type === 'chart') {
     // chart
-  } else if (type === 'filter') {
+  } else if (type === 'autofilter') {
     // filter
+    autofilter.call(this);
   } else if (type === 'freeze') {
     if (value) {
       const { ri, ci } = data.selector;
@@ -408,15 +466,22 @@ function toolbarChange(type, value) {
     }
   } else {
     data.setSelectedCellAttr(type, value);
-    if (type === 'formula') {
+    if (type === 'formula' && !data.selector.multiple()) {
       editorSet.call(this);
     }
     sheetReset.call(this);
   }
 }
 
+function sortFilterChange(ci, order, operator, value) {
+  // console.log('sort:', sortDesc, operator, value);
+  this.data.setAutoFilter(ci, order, operator, value);
+  sheetReset.call(this);
+}
+
 function sheetInitEvents() {
   const {
+    selector,
     overlayerEl,
     rowResizer,
     colResizer,
@@ -427,6 +492,7 @@ function sheetInitEvents() {
     data,
     toolbar,
     modalValidation,
+    sortFilter,
   } = this;
   // overlayer
   overlayerEl
@@ -449,9 +515,20 @@ function sheetInitEvents() {
         editor.clear();
         overlayerMousedown.call(this, evt);
       }
-    }).on('mousewheel.stop', (evt) => {
+    })
+    .on('mousewheel.stop', (evt) => {
       overlayerMousescroll.call(this, evt);
+    })
+    .on('mouseout', (evt) => {
+      const { offsetX, offsetY } = evt;
+      if (offsetY <= 0) colResizer.hide();
+      if (offsetX <= 0) rowResizer.hide();
     });
+
+  selector.inputChange = (v) => {
+    dataSetCellText.call(this, v, 'input');
+    editorSet.call(this);
+  };
 
   // slide on mobile
   bindTouch(overlayerEl.el, {
@@ -462,6 +539,9 @@ function sheetInitEvents() {
 
   // toolbar change
   toolbar.change = (type, value) => toolbarChange.call(this, type, value);
+
+  // sort filter ok
+  sortFilter.ok = (ci, order, o, v) => sortFilterChange.call(this, ci, order, o, v);
 
   // resizer finished callback
   rowResizer.finishedFn = (cRect, distance) => {
@@ -517,6 +597,12 @@ function sheetInitEvents() {
     this.focusing = overlayerEl.contains(evt.target);
   });
 
+  bind(window, 'paste', (evt) => {
+    const cdata = evt.clipboardData.getData('text/plain');
+    this.data.pasteFromText(cdata);
+    sheetReset.call(this);
+  });
+
   // for selector
   bind(window, 'keydown', (evt) => {
     if (!this.focusing) return;
@@ -559,7 +645,7 @@ function sheetInitEvents() {
         case 86:
           // ctrl + v
           paste.call(this, what);
-          evt.preventDefault();
+          // evt.preventDefault();
           break;
         case 37:
           // ctrl + left
@@ -588,11 +674,11 @@ function sheetInitEvents() {
           break;
         case 66:
           // ctrl + B
-          toolbar.trigger('font-bold');
+          toolbar.trigger('bold');
           break;
         case 73:
           // ctrl + I
-          toolbar.trigger('font-italic');
+          toolbar.trigger('italic');
           break;
         default:
           break;
@@ -634,6 +720,13 @@ function sheetInitEvents() {
           evt.preventDefault();
           break;
         case 13: // enter
+          if (altKey) {
+            const c = data.getSelectedCell();
+            const ntxt = c.text || '';
+            dataSetCellText.call(this, `${ntxt}\n`, 'input');
+            editorSet.call(this);
+            break;
+          }
           editor.clear();
           // shift + enter => move up
           // enter => move down
@@ -671,7 +764,8 @@ export default class Sheet {
     const { view, showToolbar, showContextmenu } = data.settings;
     this.el = h('div', `${cssPrefix}-sheet`);
     this.toolbar = new Toolbar(data, view.width, !showToolbar);
-    targetEl.children(this.toolbar.el, this.el);
+    this.print = new Print(data);
+    targetEl.children(this.toolbar.el, this.el, this.print.el);
     this.data = data;
     // table
     this.tableEl = h('canvas', `${cssPrefix}-table`);
@@ -700,6 +794,8 @@ export default class Sheet {
       );
     this.overlayerEl = h('div', `${cssPrefix}-overlayer`)
       .child(this.overlayerCEl);
+    // sortFilter
+    this.sortFilter = new SortFilter();
     // root element
     this.el.children(
       this.tableEl,
@@ -710,6 +806,7 @@ export default class Sheet {
       this.horizontalScrollbar.el,
       this.contextMenu.el,
       this.modalValidation.el,
+      this.sortFilter.el,
     );
     // table
     this.table = new Table(this.tableEl.el, data);
@@ -717,6 +814,14 @@ export default class Sheet {
     sheetReset.call(this);
     // init selector [0, 0]
     selectorSet.call(this, false, 0, 0);
+  }
+
+  resetData(data) {
+    this.data = data;
+    this.toolbar.resetData(data);
+    this.print.resetData(data);
+    this.selector.resetData(data);
+    this.table.resetData(data);
   }
 
   loadData(data) {
